+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Changed
 
 ### Removed
+- PHP 7.1 support
 
 
 ## [6.0.0-alpha.1]